Make lib loading more atomic
