+++ spec/ip_spec.rb
@@ -1,4 +1,7 @@
 require 'spec_helper'
+require 'packetfu/protos/eth'
+require 'packetfu/protos/ip'
+require 'packetfu/pcap'
 require 'tempfile'
 
 include PacketFu
