+++ spec/spec_helper.rb
@@ -1,7 +1,3 @@
-
-$:.unshift File.join(File.expand_path(File.dirname(__FILE__)), "..", "lib")
-require 'packetfu'
-
 puts "rspec #{RSpec::Core::Version::STRING}"
 if RSpec::Core::Version::STRING[0] == '3'
   require 'rspec/its'
@@ -13,32 +9,4 @@ if RSpec::Core::Version::STRING[0] == '3'
   end
 end
 
-
-module FakePacket
-  def layer
-    7
-  end
-end
-
-class PacketFu::FooPacket < PacketFu::Packet
-  extend FakePacket
-end
-
-class PacketFu::BarPacket < PacketFu::Packet
-  extend FakePacket
-end
-
-class PacketBaz
-end
-
-def add_fake_packets
-  PacketFu.add_packet_class(PacketFu::FooPacket)
-  PacketFu.add_packet_class(PacketFu::BarPacket)
-end
-
-def remove_fake_packets
-  PacketFu.remove_packet_class(PacketFu::FooPacket)
-  PacketFu.remove_packet_class(PacketFu::BarPacket)
-end
-
-remove_fake_packets
+require 'packetfu/common'
