+++ test/test_udp.rb
@@ -1,6 +1,5 @@
 #!/usr/bin/env ruby
 # -*- coding: binary -*-
-
 require 'test/unit'
 $:.unshift File.join(File.expand_path(File.dirname(__FILE__)), "..", "lib")
 require 'packetfu'
